+++ packages/rescript-relay/rescript-relay-ppx/library/RescriptRelayPpxLibrary.ml
@@ -6,13 +6,16 @@ let endsWithRegexp = Str.regexp ".*Resolver$"
 let commonExtension =
   Extension.declare "relay" Extension.Context.module_expr
     (let open Ast_pattern in
-    single_expr_payload (estring __))
+     single_expr_payload (estring __))
     (fun ~loc ~path:_ operationStr ->
       let op = extractGraphQLOperation ~loc operationStr in
       match op with
       | Graphql_parser.Fragment {name = fragmentName; selection_set} ->
         if Str.string_match endsWithRegexp fragmentName 0 then
           RelayResolverFragment.make ~loc
+            ~moduleName:(op |> extractTheFragmentName ~loc)
+        else if Util.fragmentIsUpdatable op then
+          UpdatableFragment.make ~loc
             ~moduleName:(op |> extractTheFragmentName ~loc)
         else
           let refetchableQueryName =
@@ -25,10 +28,13 @@ let commonExtension =
             ~hasInlineDirective:(op |> fragmentHasInlineDirective ~loc)
             ~loc
       | Operation {optype = Query} ->
-        Query.make
-          ~moduleName:(op |> extractTheQueryName ~loc)
-          ~hasRawResponseType:(op |> queryHasRawResponseTypeDirective ~loc)
-          ~loc
+        if Util.queryIsUpdatable op then
+          UpdatableQuery.make ~loc ~moduleName:(op |> extractTheQueryName ~loc)
+        else
+          Query.make
+            ~moduleName:(op |> extractTheQueryName ~loc)
+            ~hasRawResponseType:(op |> queryHasRawResponseTypeDirective ~loc)
+            ~loc
       | Operation {optype = Mutation} ->
         Mutation.make ~moduleName:(op |> extractTheMutationName ~loc) ~loc
       | Operation {optype = Subscription} ->