+++ packages/rescript-relay/rescript-relay-ppx/library/Util.ml
@@ -86,6 +86,20 @@ let queryHasRawResponseTypeDirective ~loc op =
     |> List.exists (fun (directive : Graphql_parser.directive) ->
            directive.name = "raw_response_type")
   | _ -> false
+let fragmentIsUpdatable op =
+  match op with
+  | Graphql_parser.Fragment {directives} ->
+    directives
+    |> List.exists (fun (directive : Graphql_parser.directive) ->
+           directive.name = "updatable")
+  | _ -> false
+let queryIsUpdatable op =
+  match op with
+  | Graphql_parser.Operation {optype = Query; directives} ->
+    directives
+    |> List.exists (fun (directive : Graphql_parser.directive) ->
+           directive.name = "updatable")
+  | _ -> false
 type connectionConfig = {key: string}
 let extractFragmentConnectionInfo ~loc op =
   match op with